requirements.txt
langchain==0.3.15
langchain_community==0.3.15
langchain_core==0.3.31
langchain_openai==0.3.1
langgraph==0.2.65
pandas==2.2.3
python-dotenv==1.0.1
SQLAlchemy==2.0.34
sqlparse==0.5.3
streamlit==1.40.1
streamlit_authenticator==0.4.1
